Trait pallet_custom_signatures::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Call: Parameter + UnfilteredDispatchable<Origin = Self::Origin> + GetDispatchInfo;
    type Signature: Parameter + Verify<Signer = Self::Signer> + TryFrom<Vec<u8>>;
    type Signer: IdentifyAccount<AccountId = Self::AccountId>;
    type Currency: Currency<Self::AccountId>;
    type OnChargeTransaction: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>;
    type CallFee: Get<BalanceOf<Self>>;
    type CallMagicNumber: Get<u16>;
    type UnsignedPriority: Get<TransactionPriority>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

The overarching event type.

type Call: Parameter + UnfilteredDispatchable<Origin = Self::Origin> + GetDispatchInfo

A signable call.

type Signature: Parameter + Verify<Signer = Self::Signer> + TryFrom<Vec<u8>>

User defined signature type.

type Signer: IdentifyAccount<AccountId = Self::AccountId>

User defined signer type.

type Currency: Currency<Self::AccountId>

The currency trait.

type OnChargeTransaction: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>

The call fee destination.

type CallFee: Get<BalanceOf<Self>>

The call processing fee amount.

type CallMagicNumber: Get<u16>

The call magic number.

type UnsignedPriority: Get<TransactionPriority>

A configuration for base priority of unsigned transactions.

This is exposed so that it can be tuned for particular runtime, when multiple pallets send unsigned transactions.