Trait pallet_custom_signatures::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + Into<Self::Event>;
    type Call: Parameter + UnfilteredDispatchable<Origin = Self::Origin> + GetDispatchInfo;
    type Signature: Parameter + Verify<Signer = Self::Signer> + TryFrom<Vec<u8>>;
    type Signer: IdentifyAccount<AccountId = Self::AccountId>;
    type UnsignedPriority: Get<TransactionPriority>;
}

The module’s configuration trait.

Associated Types

type Event: From<Event<Self>> + Into<Self::Event>

The overarching event type.

type Call: Parameter + UnfilteredDispatchable<Origin = Self::Origin> + GetDispatchInfo

A signable call.

type Signature: Parameter + Verify<Signer = Self::Signer> + TryFrom<Vec<u8>>

User defined signature type.

type Signer: IdentifyAccount<AccountId = Self::AccountId>

User defined signer type.

type UnsignedPriority: Get<TransactionPriority>

A configuration for base priority of unsigned transactions.

This is exposed so that it can be tuned for particular runtime, when multiple pallets send unsigned transactions.

Loading content...